requirements.txt
asgiref==3.7.2
blessed==1.19.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==5.0.1
django-jalali-date==1.1.0
fire==0.4.0
halo==0.0.31
idna==3.2
jdatetime==4.1.1
log-symbols==0.0.14
numpy==1.26.4
pandas==2.2.0
pathspec==0.9.0
Pygments==2.10.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
readchar==2.0.1
regex==2023.10.3
requests==2.26.0
requests-toolbelt==0.9.1
runflare==1.1.8
scheduler==0.8.5
six==1.16.0
spinners==0.0.24
sqlparse==0.4.4
termcolor==1.1.0
typeguard==4.1.5
typing_extensions==4.9.0
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
wcwidth==0.2.5
websockets==10.0
